Switch to servant 0.19 and NamedRoutes
Neat new feature. However, the way to be able to use it was a bit more
bumpy than I was hoping for. I'm now using the nightly stackage
resolver, and I use sytest and related dependencies in a version that's
currently only existing in PRs, waiting for dependencies to ship the
aeson 2 compatiblity. Since this is a toy project, I don't have a real
problem using these. I was interested in the named routes and so far I'm
really pleased with them. Let's see how that goes in the future.

I deleted the Hib.Docs module because it got in the way while upgrading
and I don't see enough benefit to keep it around. I learned enough while
giving it a try initially.

See:

* https://www.tweag.io/blog/2022-02-24-named-routes/